Added PREMUL and qt_div_255 here since we don't want to rely on a private header. It has changed in 5.5 breaking compilation.
